ref(metrics): move metrics api metrics details endpoint to separate file
