+++ src/sentry/api/endpoints/organization_metrics.py
@@ -2,19 +2,9 @@ from collections.abc import Sequence
 
 from rest_framework.exceptions import ParseError
 from rest_framework.request import Request
-from rest_framework.response import Response
 
-from sentry.api.api_owners import ApiOwner
-from sentry.api.api_publish_status import ApiPublishStatus
-from sentry.api.base import region_silo_endpoint
-from sentry.api.bases.organization import (
-    OrganizationAndStaffPermission,
-    OrganizationEndpoint,
-    OrganizationPermission,
-)
+from sentry.api.bases.organization import OrganizationPermission
 from sentry.auth.elevated_mode import has_elevated_mode
-from sentry.exceptions import InvalidParams
-from sentry.sentry_metrics.querying.metadata import get_metrics_meta
 from sentry.sentry_metrics.use_case_id_registry import (
     UseCaseID,
     UseCaseIDAPIAccess,
@@ -98,25 +88,3 @@ def get_use_case_ids(request: Request) -> Sequence[UseCaseID]:
 
 class OrganizationMetricsEnrollPermission(OrganizationPermission):
     scope_map = {"PUT": ["org:read", "org:write", "org:admin"]}
-
-
-@region_silo_endpoint
-class OrganizationMetricsDetailsEndpoint(OrganizationEndpoint):
-    publish_status = {
-        "GET": ApiPublishStatus.EXPERIMENTAL,
-    }
-    owner = ApiOwner.TELEMETRY_EXPERIENCE
-    permission_classes = (OrganizationAndStaffPermission,)
-
-    """Get the metadata of all the stored metrics including metric name, available operations and metric unit"""
-
-    def get(self, request: Request, organization) -> Response:
-        projects = self.get_projects(request, organization)
-        if not projects:
-            raise InvalidParams("You must supply at least one project to see its metrics")
-
-        metrics = get_metrics_meta(
-            organization=organization, projects=projects, use_case_ids=get_use_case_ids(request)
-        )
-
-        return Response(metrics, status=200)
